refactor(frontend/copilot): split UsageLimits into UsagePopover and UsageLimitReachedCard
Extract two fully independent components from the previous shared
UsagePanelContent abstraction. UsageLimitReachedCard now always renders
the "Go to billing" CTA — gating on whether the limit is reached lives
in ChatContainer, so the card itself can stay simple.

Reorganises the folder under CONTRIBUTING.md conventions
(ComponentName/ComponentName.tsx + useComponentName.ts), pulls UsageBar
and StorageBar out as small render primitives reused by the credits
page, and moves formatBytes into usageHelpers.ts. Stale tests for the
removed UsagePanelContent are deleted; both new components get focused
integration tests.
