+++ autogpt_platform/frontend/src/app/(platform)/copilot/components/UsageLimits/UsagePopover/useUsagePopover.ts
@@ -1,0 +1,14 @@
+import { useGetV2GetCopilotUsage } from "@/app/api/__generated__/endpoints/chat/chat";
+import type { CoPilotUsagePublic } from "@/app/api/__generated__/models/coPilotUsagePublic";
+
+export function useUsagePopover() {
+  const { data: usage, isSuccess } = useGetV2GetCopilotUsage({
+    query: {
+      select: (res) => res.data as CoPilotUsagePublic,
+      refetchInterval: 30000,
+      staleTime: 10000,
+    },
+  });
+
+  return { usage, isSuccess };
+}
