+++ autogpt_platform/frontend/src/app/(platform)/copilot/components/usageHelpers.ts
@@ -32,3 +32,19 @@ export function formatResetTime(
     timeZoneName: "short",
   });
 }
+
+export function formatBytes(bytes: number): string {
+  const KB = 1024;
+  const MB = KB * 1024;
+  const GB = MB * 1024;
+  if (bytes < KB) return `${bytes} B`;
+  if (bytes < MB) {
+    const kb = Math.round(bytes / KB);
+    return kb >= 1024 ? `${(bytes / MB).toFixed(1)} MB` : `${kb} KB`;
+  }
+  if (bytes < GB) {
+    const mb = Math.round(bytes / MB);
+    return mb >= 1024 ? `${(bytes / GB).toFixed(1)} GB` : `${mb} MB`;
+  }
+  return `${(bytes / GB).toFixed(1)} GB`;
+}
